fix: fixing carousel top article
